extract FeedbackButtons and remove indirection
