+++ vscode/webviews/chat/components/FeedbackButtons.module.css
@@ -1,0 +1,20 @@
+.feedback-buttons {
+    display: flex;
+    flex-direction: row;
+    gap: 0.15rem;
+}
+
+.feedback-button[disabled] {
+    /* VSCodeButton's default cursor is not-allowed, but that's different to
+       native VS Code and feels off, especially when it quickly changes to
+       not-allowed after you submit feedback. So we reset it back to the
+       default cursor to fit in nicer with standard VS Code native behaviour */
+    cursor: default;
+}
+
+.thumbs-down-feedback-container {
+    display: flex;
+    align-items: center;
+    gap: calc(var(--spacing) / 4);
+}
+
