+++ vscode/webviews/chat/components/FeedbackButtons.tsx
@@ -1,0 +1,84 @@
+import { VSCodeButton, VSCodeLink } from '@vscode/webview-ui-toolkit/react'
+import classNames from 'classnames'
+import { useCallback, useState } from 'react'
+import { CODY_FEEDBACK_URL } from '../../../src/chat/protocol'
+import styles from './FeedbackButtons.module.css'
+
+export interface FeedbackButtonsProps {
+    className: string
+    disabled?: boolean
+    feedbackButtonsOnSubmit: (text: string) => void
+}
+
+export const FeedbackButtons: React.FunctionComponent<FeedbackButtonsProps> = ({
+    className,
+    feedbackButtonsOnSubmit,
+}) => {
+    const [feedbackSubmitted, setFeedbackSubmitted] = useState('')
+
+    const onFeedbackBtnSubmit = useCallback(
+        (text: string) => {
+            feedbackButtonsOnSubmit(text)
+            setFeedbackSubmitted(text)
+        },
+        [feedbackButtonsOnSubmit]
+    )
+
+    return (
+        <div className={classNames(styles.feedbackButtons, className)}>
+            {!feedbackSubmitted && (
+                <>
+                    <VSCodeButton
+                        className={classNames(styles.feedbackButton)}
+                        appearance="icon"
+                        type="button"
+                        onClick={() => onFeedbackBtnSubmit('thumbsUp')}
+                    >
+                        <i className="codicon codicon-thumbsup" />
+                    </VSCodeButton>
+                    <VSCodeButton
+                        className={classNames(styles.feedbackButton)}
+                        appearance="icon"
+                        type="button"
+                        onClick={() => onFeedbackBtnSubmit('thumbsDown')}
+                    >
+                        <i className="codicon codicon-thumbsdown" />
+                    </VSCodeButton>
+                </>
+            )}
+            {feedbackSubmitted === 'thumbsUp' && (
+                <VSCodeButton
+                    className={classNames(styles.feedbackButton)}
+                    appearance="icon"
+                    type="button"
+                    disabled={true}
+                    title="Thanks for your feedback"
+                >
+                    <i className="codicon codicon-thumbsup" />
+                    <i className="codicon codicon-check" />
+                </VSCodeButton>
+            )}
+            {feedbackSubmitted === 'thumbsDown' && (
+                <span className={styles.thumbsDownFeedbackContainer}>
+                    <VSCodeButton
+                        className={classNames(styles.feedbackButton)}
+                        appearance="icon"
+                        type="button"
+                        disabled={true}
+                        title="Thanks for your feedback"
+                    >
+                        <i className="codicon codicon-thumbsdown" />
+                        <i className="codicon codicon-check" />
+                    </VSCodeButton>
+                    <VSCodeLink
+                        href={String(CODY_FEEDBACK_URL)}
+                        target="_blank"
+                        title="Help improve Cody by providing more feedback about the quality of this response"
+                    >
+                        Give Feedback
+                    </VSCodeLink>
+                </span>
+            )}
+        </div>
+    )
+}
